Commit all

+++ src/main/java/peaksoft/dto/request/RestaurantRequest.java
@@ -1,11 +1,9 @@
 package peaksoft.dto.request;
 
-import java.math.BigDecimal;
-
 public record RestaurantRequest(
         String name,
         String location,
         String restType,
-        BigDecimal service
+        int service
 ) {
 }
